+++ build.gradle
@@ -16,7 +16,7 @@ repositories {
 }
 
 allprojects {
-	group 'kr.co.yapp'
+	group 'kr.co.knowledgerally'
 	version '0.0.1-SNAPSHOT'
 }
