+++ knowlly-api/src/main/java/kr/co/knowledgerally/api/KnowllyApplication.java
@@ -1,4 +1,4 @@
-package kr.co.yapp.api;
+package kr.co.knowledgerally.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
