+++ knowlly-core/src/main/java/kr/co/knowledgerally/core/lecture/entity/Lecture.java
@@ -1,5 +1,6 @@
-package kr.co.yapp.core.lecture.entity;
+package kr.co.knowledgerally.core.lecture.entity;
 
+import kr.co.knowledgerally.core.user.entity.Coach;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
@@ -12,8 +13,8 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-@Table(name = "lecture_image")
-public class LectureImage {
+@Table(name = "lecture")
+public class Lecture {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
@@ -21,11 +22,26 @@ public class LectureImage {
 
     @MapsId("id")
     @ManyToOne
-    @JoinColumn(name = "lecture_id")
-    private Lecture lecture;
+    @JoinColumn(name = "coach_id")
+    private Coach coach;
+
+    @MapsId("id")
+    @ManyToOne
+    @JoinColumn(name = "category_id")
+    private Category category;
 
     @Column(nullable = false)
-    private String lectureImgUrl;
+    private String title;
+
+    @Column(nullable = false)
+    private String introduce;
+
+    @Column(nullable = false)
+    private int price;
+
+    @Builder.Default
+    @Column(nullable = false)
+    private Boolean isActive = true;
 
     @Builder.Default
     @CreationTimestamp
